requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
pillow==8.3.2
numpy==1.21.2
onnxruntime==1.9.0
torch==1.9.0
torchvision==0.10.0
pytest==6.2.5
requests==2.26.0
python-dotenv==0.19.0 